Use DiffUtils for patching
Unfortunately, native code is nessicarry for reasonable performance
when calculating large diffs.
Although the plain python engine seems to work on the currently small
diffs, benchmarks indicate it can be up to 10 times slower on large
ones.
However, precompiled wheels are available for linux and windows,
and are automatically downloaded by the scripts.
Blame apple for the lack of mac builds, since I'm not paying $1000
dollars to compile some software.

Fixes #1